backend/requirements.txt
# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3

# Message Queue
aio-pika==9.4.0

# ML
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Utils
python-dotenv==1.0.0
httpx==0.26.0
